support uint16 pose ids

+++ src/core/bluetooth_communication/debug_capture/position_capture_test_pipeline.cpp
@@ -49,7 +49,7 @@ namespace
         std::int32_t heading_urad = 0;
         std::uint16_t confidence_position = 0U;
         std::uint16_t confidence_heading = 0U;
-        std::uint8_t pose_id = 0U;
+        std::uint16_t pose_id = 0U;
         std::uint8_t branch_id = 0U;
         std::uint8_t flags = 0U;
     };
