+++ src/core/motion_mcu_communication/outgoing_payloads/service/position_correction_payload.hpp
@@ -4,5 +4,5 @@
 
 namespace position_correction_payload
 {
-    bool send(std::uint8_t pose_id, std::uint8_t branch_id);
+    bool send(std::uint16_t pose_id, std::uint8_t branch_id);
 }
