+++ src/core/position_sensorfusion/filtered_global_offset_fusion/filtered_global_offset_fusion.hpp
@@ -18,7 +18,7 @@ namespace filtered_global_offset_fusion
         std::uint16_t confidence_position = 0U;
         std::uint16_t confidence_heading = 0U;
 
-        std::uint8_t pose_id = 0U;
+        std::uint16_t pose_id = 0U;
         std::uint8_t branch_id = 0U;
 
         bool has_offset = false;
